disabled videocompression and check if video exists (#16044)

+++ frontend/cypress.config.ts
@@ -8,6 +8,7 @@ export default defineConfig({
   trashAssetsBeforeRuns: true,
   videosFolder: './gui_test_screenshots/cypress/videos',
   video: true,
+  videoCompression: false,
   reporter: './node_modules/cypress-multi-reporters',
   reporterOptions: {
     configFile: 'reporter-config.json',
@@ -92,7 +93,7 @@ export default defineConfig({
             const failures = results.tests.some((test) =>
               test.attempts.some((attempt) => attempt.state === 'failed')
             )
-            if (!failures) {
+            if (!failures && fs.existsSync(results.video)) {
               // delete the video if the spec passed and no tests retried
               fs.unlinkSync(results.video)
             }
